Feature/10.compartilhando recomendacoes

+++ AluGames/src/main/kotlin/br/com/luizlemetech/alugames/modelo/Plano.kt
@@ -1,8 +1,10 @@
 package br.com.luizlemetech.alugames.modelo
 
+import java.math.BigDecimal
+
 abstract class Plano(val tipo: String) {
 
-    open fun obterValor(aluguel: Aluguel): Double {
-        return aluguel.jogo.preco * aluguel.periodo.emDias
+    open fun obterValor(aluguel: Aluguel): BigDecimal {
+        return aluguel.jogo.preco * aluguel.periodo.emDias.toBigDecimal()
     }
 }
